Update migration and models for Employee and Product
Renamed the migration class `InitialCreate` to `Martify`.
Added an optional `ImagePath` property to the `Employee` entity.
Updated the `ImagePath` property in the `Product` entity to be optional.
Added a required `Phone` property to the `Employee` entity with a max length of 15.
Updated the `Martify.db` file to reflect schema changes.
Adjusted migration files and model snapshots to align with these updates.